+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.0.1
 
+- Added ClamAV settings for virus detection, see example configuration file
 - Switched to Perun as authentication. The example configuration file now contains the values for perun. Set "de.samply.development" to true
   to enable the development mode, which is necessary for now.
 - There is only one configuration file: `bbmri.negotiator.xml` which includes all previously existing configuration settings.